# docs/swagger.yaml
definitions:
  service.UserInput:
    properties:
      name:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /like/addLike:
    post:
      parameters:
      - description: 用戶ID
        in: formData
        name: userId
        type: string
      - description: 貼文ID
        in: formData
        name: postId
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 貼文按讚
      tags:
      - 貼文按讚
  /ping:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首頁
  /post/createPost:
    post:
      parameters:
      - description: 用戶名
        in: formData
        name: author
        type: string
      - description: 標題
        in: formData
        name: title
        type: string
      - description: 內容
        in: formData
        name: content
        type: string
      - description: 照片
        in: formData
        name: image
        type: file
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 創建貼文
      tags:
      - 貼文資料
  /post/getPostList:
    get:
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 貼文列表
      tags:
      - 貼文資料
  /user/createUser:
    post:
      parameters:
      - description: 電子郵件
        in: formData
        name: email
        required: true
        type: string
      - description: 用戶名
        in: formData
        name: name
        required: true
        type: string
      - description: 密碼
        in: formData
        name: password
        required: true
        type: string
      - description: 確認密碼
        in: formData
        name: repassword
        required: true
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 新增用戶
      tags:
      - 用戶資料
  /user/deleteUser:
    get:
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 刪除用戶
      tags:
      - 用戶資料
  /user/findUserByNameAndPwd:
    post:
      parameters:
      - description: 用戶名和密碼
        in: body
        name: UserInput
        required: true
        schema:
          $ref: '#/definitions/service.UserInput'
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 用戶登入
      tags:
      - 用戶資料
  /user/getUserList:
    get:
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - 用戶資料
  /user/updateUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: 用戶名
        in: formData
        name: name
        type: string
      - description: 密碼
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 修改用戶
      tags:
      - 用戶資料
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"